requirements.txt
streamlit==1.25.0
google-api-python-client==2.95.0
google-auth==2.22.0
pandas==1.5.3
